feat(rds): removing ihe stack
Refs: #1923
