+++ .github/workflows/_deploy-cdk.yml
@@ -24,10 +24,6 @@ on:
         required: true
         type: string
         description: "the name of the CDK stack we're deploying"
-      ihe_stack:
-        required: false
-        type: string
-        description: "the name of the IHE CDK stack we're deploying"
       add_sentry_release:
         description: "the name of the CDK stack we're deploying"
         required: false
@@ -230,35 +226,6 @@ jobs:
           cdk_action: "deploy --verbose --require-approval never"
           cdk_version: "2.122.0"
           cdk_stack: "${{ inputs.cdk_stack }}"
-          cdk_env: "${{ inputs.deploy_env }}"
-        env:
-          INPUT_PATH: "metriport/packages/infra"
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
-          METRIPORT_VERSION: ${{ github.sha }}
-
-      # IHE Stack 
-      - name: Diff IHE Stack CDK
-        uses: metriport/deploy-with-cdk@master
-        if: inputs.ihe_stack != null
-        with:
-          cdk_action: "diff"
-          cdk_version: "2.122.0"
-          cdk_stack: "${{ inputs.ihe_stack }}"
-          cdk_env: "${{ inputs.deploy_env }}"
-        env:
-          INPUT_PATH: "metriport/packages/infra"
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ inputs.AWS_REGION }}
-      - name: Deploy IHE Stack CDK
-        uses: metriport/deploy-with-cdk@master
-        if: inputs.ihe_stack != null
-        with:
-          cdk_action: "deploy --verbose --require-approval never"
-          cdk_version: "2.122.0"
-          cdk_stack: "${{ inputs.ihe_stack }}"
           cdk_env: "${{ inputs.deploy_env }}"
         env:
           INPUT_PATH: "metriport/packages/infra"
